moved config options and settings classes to independent files
